HTBHF-2230 Added claim creator utility (WIP).
