+++ .gitignore
@@ -27,5 +27,3 @@ gradle.properties
 api/out/
 db/out/
 manifest/out/
-claimant-creator/out
-claimant-creator/src/main/resources/claimant.yml
