+++ api/src/test/java/uk/gov/dhsc/htbhf/claimant/creator/ClaimantCreator.java
@@ -1,0 +1,31 @@
+package uk.gov.dhsc.htbhf.claimant.creator;
+
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.fasterxml.jackson.dataformat.yaml.YAMLFactory;
+import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
+import com.fasterxml.jackson.datatype.jsr310.ser.LocalDateTimeSerializer;
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Profile;
+
+import java.time.LocalDateTime;
+import java.time.format.DateTimeFormatter;
+
+@SpringBootApplication
+@Profile("test-claimant-creator")
+public class ClaimantCreator {
+
+    @Bean
+    public ObjectMapper objectMapper() {
+        ObjectMapper objectMapper = new ObjectMapper(new YAMLFactory());
+        JavaTimeModule javaTimeModule = new JavaTimeModule();
+        javaTimeModule.addSerializer(LocalDateTime.class, new LocalDateTimeSerializer(DateTimeFormatter.ISO_LOCAL_DATE_TIME));
+        objectMapper.findAndRegisterModules();
+        return objectMapper;
+    }
+
+    public static void main(String[] args) {
+        SpringApplication.run(ClaimantCreator.class, args);
+    }
+}
