+++ api/src/test/java/uk/gov/dhsc/htbhf/claimant/creator/dwp/entities/Household.java
@@ -1,0 +1,10 @@
+package uk.gov.dhsc.htbhf.claimant.creator.dwp.entities;
+
+import java.util.Set;
+
+public interface Household {
+
+    String getHouseholdIdentifier();
+
+    Set<? extends Child> getChildren();
+}
