+++ api/src/test/java/uk/gov/dhsc/htbhf/claimant/creator/dwp/respository/UCHouseholdRepository.java
@@ -1,0 +1,12 @@
+package uk.gov.dhsc.htbhf.claimant.creator.dwp.respository;
+
+import org.springframework.data.repository.CrudRepository;
+import uk.gov.dhsc.htbhf.claimant.creator.dwp.entities.uc.UCHousehold;
+
+import java.util.UUID;
+
+/**
+ * JPA gov.uk.dhsc.claimant.creator.repository for Universal Credit households (members of the household are persisted via the household).
+ */
+public interface UCHouseholdRepository extends CrudRepository<UCHousehold, UUID> {
+}
