+++ api/src/test/resources/claimant-creator/claimant-example.yml
@@ -1,0 +1,19 @@
+# all fields are mandatory unless stated otherwise
+claimant:
+  first-name: james
+  last-name: smith
+  nino: EE123456A
+  date-of-birth: 01/01/1970
+  mobile: 07123456789
+  email-address: test@domain.com
+  address:
+    first-line: 1 Street
+    second-line: Fake building # optional
+    townOrCity: London
+    county: Greator London # optional
+    postcode: BS11AA
+  children-age: # optional
+    - age: P3Y11M24D # duration in ISO-8601 format
+      at: START_OF_FIRST_CYCLE
+    - age: P6M
+      at: START_OF_NEXT_CYCLE
