+++ api/src/test/resources/claimant-creator/claimant.yml
@@ -1,0 +1,19 @@
+# all fields are mandatory unless stated otherwise
+firstName: james
+lastName: smith
+nino: EE123456A
+dateOfBirth: 1970-01-01
+expectedDeliveryDate: 2020-01-01 # TODO make this a duration
+mobile: 07123456789
+emailAddress: test@domain.com
+address:
+  addressLine1: 1 Street
+  addressLine2: Fake building # optional
+  townOrCity: London
+  county: Greator London # optional
+  postcode: BS11AA
+childrenAgeInfo: # optional
+  - age: P3Y11M24D # duration in ISO-8601 format
+    at: START_OF_FIRST_CYCLE
+  - age: P6M
+    at: START_OF_NEXT_CYCLE
