+++ settings.gradle
@@ -1,8 +1,6 @@
 rootProject.name = 'htbhf-claimant-service-parent'
-include "htbhf-claimant-service", "htbhf-claimant-service-db", "htbhf-claimant-service-manifest", 'htbhf-claimant-creator'
+include "htbhf-claimant-service", "htbhf-claimant-service-db", "htbhf-claimant-service-manifest"
 
 project(":htbhf-claimant-service").projectDir = file("api")
 project(":htbhf-claimant-service-db").projectDir = file("db")
 project(":htbhf-claimant-service-manifest").projectDir = file("manifest")
-project(":htbhf-claimant-creator").projectDir = file("test-claimant-creator")
-
