--- test-claimant-creator/lombok.config
@@ -1,2 +1,0 @@
-lombok.anyConstructor.addConstructorProperties=true
-lombok.addLombokGeneratedAnnotation=true
