--- test-claimant-creator/src/main/java/uk/gov/dhsc/htbhf/claimant/creator/entities/claimant/EligibilityStatus.java
@@ -1,16 +1,0 @@
-package uk.gov.dhsc.htbhf.claimant.creator.entities.claimant;
-
-import com.fasterxml.jackson.annotation.JsonFormat;
-
-/**
- * The possible values for the eligibility of a claimant.
- */
-@JsonFormat(shape = JsonFormat.Shape.STRING)
-public enum EligibilityStatus {
-    ELIGIBLE,
-    INELIGIBLE,
-    PENDING,
-    NO_MATCH,
-    ERROR,
-    DUPLICATE
-}
