--- test-claimant-creator/src/main/java/uk/gov/dhsc/htbhf/claimant/creator/entities/claimant/PaymentStatus.java
@@ -1,10 +1,0 @@
-package uk.gov.dhsc.htbhf.claimant.creator.entities.claimant;
-
-/**
- * Status of a payment.
- */
-public enum PaymentStatus {
-    SUCCESS,
-    FAILURE,
-    ERROR
-}
