--- test-claimant-creator/src/main/java/uk/gov/dhsc/htbhf/claimant/creator/entities/dwp/Household.java
@@ -1,10 +1,0 @@
-package uk.gov.dhsc.htbhf.claimant.creator.entities.dwp;
-
-import java.util.Set;
-
-public interface Household {
-
-    String getHouseholdIdentifier();
-
-    Set<? extends Child> getChildren();
-}
