--- test-claimant-creator/src/main/resources/claimant-example.yml
@@ -1,19 +1,0 @@
-# all fields are mandatory unless stated otherwise
-firstName: David
-lastName: Test
-nino: EE222222A
-dateOfBirth: 1970-01-01
-expectedDeliveryDate: 2020-01-01
-mobile: 07123456789
-emailAddress: dwybourn@equalexperts.com
-address:
-  addressLine1: 1 Street
-  addressLine2: Fake building # optional
-  townOrCity: London
-  county: Greator London # optional
-  postcode: BS11AA
-childrenAgeInfo: # optional
-  - age: P3Y11M21D # duration in ISO-8601 format
-    at: START_OF_NEXT_CYCLE
-  - age: P6M
-    at: START_OF_FIRST_CYCLE
